Pagination refactoring
This intoduces a new pagination logic to go through the result
sets from OAI requests. It no longer iterates over the xml document
to retrieve the OAI records for building the tuples. Instead, it
now parses all records from a result set into a list right after
loading the xml response, so that they can accessed later on for
building the tuples.